Split menu bulk mutations
