+++ saleor/graphql/menu/bulk_mutations/__init__.py
@@ -1,0 +1,7 @@
+from .menu_bulk_delete import MenuBulkDelete
+from .menu_item_bulk_delete import MenuItemBulkDelete
+
+__all__ = [
+    "MenuBulkDelete",
+    "MenuItemBulkDelete",
+]
